README: Fix line endings and center Bitdeli badge
We added Bitdeli statistics to our repository so we can see how useful
we are to everyone! Yay! But let's make it look just a bit prettier by
centering it. Not sure who screwed up the line endings on this guy
either, but we'll go ahead and fix them for the future.