create_fp-lib-table: fp-lib-table creation script
This script allows the user to create an fp-lib-table file without
having to actually do any work. This is great for creating/updating the
template and/or creating/updating another fp-lib-table file. The script
works by searching a given pattern in a given module directory and
creating an fp-lib-table entry for each matching library.