mod: tssop: Center value silk under package
Because KiCad doesn't support text alignment for module text, it's not
possible to left-align silkscreen under the package. As-is, long values
produce undesirable results. Move the values to absolute center under
the package for consistent result that is visually appealing and never
allows value text to cover the pin 1 indicator.