scripts: kicad_install.sh: Don't sync existing
KiCad has decided to make our lives difficult by putting everything in a
subdirectory now. To make everything compatible, we can't just move
stuff willy nilly, so we're going to have to start tracking files within
this new subdirectory. The thing is, we don't want an install sync
overwriting these files and causing us to have to go checkout the latest
versions from the repo. Hopefully this is a suitable fix, we'll see how
it goes.